chore(auth-allowlist): Move Auth Allow list from internal to pkg
